pyproject.toml
[project]
name = "mcp-email"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
    "mcp",
    "python-dotenv",
    "beautifulsoup4",
    "clean-text",


    # Gmail API + OAuth
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_email*"]
